+++ plugins/storage/1404.go
@@ -1,16 +1,17 @@
-package dev
+package storage
 
 import (
 	"megaCrawler/crawlers"
 	"megaCrawler/extractors"
+	"strings"
 
 	"github.com/gocolly/colly/v2"
 )
 
 func init() {
-	engine := crawlers.Register("1417", "Vietnam Law & Legal Forum Magazine", "https://vietnamlawmagazine.vn")
+	engine := crawlers.Register("1404", "菲律宾每日问询者报", "http://www.inquirer.net")
 
-	engine.SetStartingURLs([]string{"https://vietnamlawmagazine.vn/sitemap.xml"})
+	engine.SetStartingURLs([]string{"https://www.inquirer.net/sitemap_index.xml"})
 
 	extractorConfig := extractors.Config{
 		Author:       true,
@@ -24,16 +25,15 @@ func init() {
 	}
 
 	extractorConfig.Apply(engine)
-
 	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
+		if strings.Contains(element.Text, ".xml") {
+			engine.Visit(element.Text, crawlers.Index)
+			return
+		}
 		engine.Visit(element.Text, crawlers.News)
 	})
 
-	engine.OnHTML(".article__sapo", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Description += element.Text
-	})
-
-	engine.OnHTML(".article__body > p > span", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+	engine.OnHTML("div.article_align > div > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
 		ctx.Content += element.Text
 	})
 }
