+++ plugins/storage/1419.go
@@ -1,4 +1,4 @@
-package dev
+package storage
 
 import (
 	"megaCrawler/crawlers"
@@ -8,9 +8,9 @@ import (
 )
 
 func init() {
-	engine := crawlers.Register("1404", "菲律宾每日问询者报", "http://www.inquirer.net")
+	engine := crawlers.Register("1419", "越南加", "https://www.vietnamplus.vn")
 
-	engine.SetStartingURLs([]string{"https://www.inquirer.net/sitemap_index.xml"})
+	engine.SetStartingURLs([]string{"https://www.vietnamplus.vn/sitemaps/google-news.xml"})
 
 	extractorConfig := extractors.Config{
 		Author:       true,
@@ -24,11 +24,16 @@ func init() {
 	}
 
 	extractorConfig.Apply(engine)
+
 	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
 		engine.Visit(element.Text, crawlers.News)
 	})
 
-	engine.OnHTML("div.article_align > div > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+	engine.OnHTML("div.article__sapo.cms-desc > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Description += element.Text
+	})
+
+	engine.OnHTML(".article__body.zce-content-body.cms-body > p", func(element *colly.HTMLElement, ctx *crawlers.Context) {
 		ctx.Content += element.Text
 	})
 }
